refactor(689): split ProviderConfigService into model-cache + compatibility mixins
Issue #689 — agent_provider_service.py had grown to 1044 lines mixing
three concerns: provider config CRUD, model-cache/live-fetch, and
compatibility probe/catalog/freshness logic.

Extract two mixins (pure cut/paste, zero logic change):
- src/services/provider_model_cache.py — ProviderModelCacheMixin +
  ProviderModelCacheEntry: load/save_model_cache, refresh_models_for_provider,
  refresh_all_models, _fetch_live_models and the per-provider _fetch_* methods.
- src/services/provider_model_compatibility.py — ProviderModelCompatibilityMixin +
  ProviderModelCompatibilityRecord: config_fingerprint, get/ensure compatibility,
  freshness checks, export_compatibility_catalog, build_compatibility_payload.

ProviderConfigService now inherits both mixins
(class ProviderConfigService(ProviderModelCacheMixin, ProviderModelCompatibilityMixin)).
Mixin inheritance — NOT composition/delegation — is deliberate: tests
monkeypatch _fetch_json / _fetch_live_models / refresh_models_for_provider on
the instance (and the class) and rely on sibling methods dispatching through
self; composition would make those patches invisible. The public API stays
byte-identical (45 attrs via MRO); dataclasses + settings-key constants are
re-exported from agent_provider_service via explicit import + __all__, so the
4 callers (web settings handlers, agent manager, CLI provider, provider_service)
and ~18 test files import unchanged.

Closes #689
